+++ scripts/install_js.sh
@@ -6,7 +6,7 @@ if ! which npm &> /dev/null; then
     sudo apt-get install npm
 fi
 
-npm --save-dev install \
+npm --prefix . --save-dev install \
     qunit \
     karma \
     karma-qunit \
@@ -17,5 +17,5 @@ npm --save-dev install \
     rollup-plugin-cleanup \
     rollup-plugin-terser \
 
-npm --no-save install \
+npm --prefix . --no-save install \
     https://github.com/jquery/jquery#main